refactor: 表格拆分模块化
