+++ .prettierrc
@@ -1,7 +1,7 @@
 {
   "singleQuote": true,
   "trailingComma": "all",
-  "printWidth": 80,
+  "printWidth": 120,
   "overrides": [
     {
       "files": ".prettierrc",
