+++ packages/client/src/components/admin-layout/Permissions/index.tsx
@@ -8,7 +8,8 @@ import { LockOutlined } from '@ant-design/icons';
 import cls from 'classnames';
 import { uid, isValid } from '@formily/shared';
 import { Resource } from '../../../resource';
-import { TableRowContext, useTable } from '../../../schemas/table';
+import { TableRowContext } from '../../../schemas/table/context';
+import { useTable } from '../../../schemas/table/hooks/useTable';
 import { useRequest } from 'ahooks';
 import { VisibleContext } from '../../../context';
 import { connect, observer, useForm } from '@formily/react';
