+++ packages/client/src/components/schema-renderer/index.tsx
@@ -52,11 +52,8 @@ import { Menu } from '../../schemas/menu';
 import { Password } from '../../schemas/password';
 import { Radio } from '../../schemas/radio';
 import { Select } from '../../schemas/select';
-import {
-  CollectionFieldContext,
-  Table,
-  TableRowContext,
-} from '../../schemas/table';
+import { Table } from '../../schemas/table';
+import { CollectionFieldContext, TableRowContext } from '../../schemas/table/context';
 import { Tabs } from '../../schemas/tabs';
 import { TimePicker } from '../../schemas/time-picker';
 import { Upload } from '../../schemas/upload';
