+++ packages/client/src/schemas/form/index.tsx
@@ -44,7 +44,7 @@ import {
 import { useResource as useGeneralResource } from '../../hooks/useResource';
 import { Resource } from '../../resource';
 import { BaseResult } from '@ahooksjs/use-request/lib/types';
-import { CollectionFieldContext } from '../table';
+import { CollectionFieldContext } from '../table/context';
 import { useTranslation } from 'react-i18next';
 
 export interface DescriptionsContextProps {
