+++ packages/client/src/schemas/grid/DesignableBar.tsx
@@ -27,7 +27,7 @@ import { useEffect } from 'react';
 import { uid } from '@formily/shared';
 import { getSchemaPath } from '../../components/schema-renderer';
 import { useCollection, useCollectionContext } from '../../constate';
-import { useTable } from '../table';
+import { useTable } from '../table/hooks/useTable';
 import { DragHandle } from '../../components/Sortable';
 import { useCompile } from '../../hooks/useCompile';
