+++ packages/client/src/schemas/table/TableActionBar.tsx
@@ -1,0 +1,106 @@
+import React, { useContext, useState } from 'react';
+import { observer } from '@formily/react';
+import cls from 'classnames';
+import { DndContext, DragOverlay } from '@dnd-kit/core';
+import { useDesignable } from '..';
+import { DisplayedMapProvider, useClient } from '../../constate';
+import { AddActionButton } from './AddActionButton';
+import { Actions } from './Actions';
+import { findPropertyByPath } from '../../components/schema-renderer';
+
+export const TableActionBar = observer((props: any) => {
+  const { align = 'top' } = props;
+  // const { schema, designable } = useDesignable();
+  const { root, schema, insertAfter, remove, appendChild } = useDesignable();
+  const moveToAfter = (path1, path2, extra = {}) => {
+    if (!path1 || !path2) {
+      return;
+    }
+    if (path1.join('.') === path2.join('.')) {
+      return;
+    }
+    const data = findPropertyByPath(root, path1);
+    if (!data) {
+      return;
+    }
+    remove(path1);
+    return insertAfter(
+      {
+        ...data.toJSON(),
+        ...extra,
+      },
+      path2,
+    );
+  };
+  const { createSchema, removeSchema, updateSchema } = useClient();
+
+  const [dragOverlayContent, setDragOverlayContent] = useState('');
+  return (
+    <DndContext
+      onDragStart={(event) => {
+        setDragOverlayContent(event.active.data?.current?.title || '');
+        // const previewRef = event.active.data?.current?.previewRef;
+        // if (previewRef) {
+        //   setDragOverlayContent(previewRef?.current?.innerHTML);
+        // } else {
+        //   setDragOverlayContent('');
+        // }
+      }}
+      onDragEnd={async (event) => {
+        const path1 = event.active?.data?.current?.path;
+        const path2 = event.over?.data?.current?.path;
+        const align = event.over?.data?.current?.align;
+        const draggable = event.over?.data?.current?.draggable;
+        if (!path1 || !path2) {
+          return;
+        }
+        if (path1.join('.') === path2.join('.')) {
+          return;
+        }
+        if (!draggable) {
+          console.log('alignalignalignalign', align);
+          const p = findPropertyByPath(root, path1);
+          if (!p) {
+            return;
+          }
+          remove(path1);
+          const data = appendChild(
+            {
+              ...p.toJSON(),
+              'x-align': align,
+            },
+            path2,
+          );
+          await updateSchema(data);
+        } else {
+          const data = moveToAfter(path1, path2, {
+            'x-align': align,
+          });
+          await updateSchema(data);
+        }
+      }}
+    >
+      <DragOverlay
+        dropAnimation={{
+          duration: 10,
+          easing: 'cubic-bezier(0.18, 0.67, 0.6, 1.22)',
+        }}
+        style={{ pointerEvents: 'none', whiteSpace: 'nowrap' }}
+      >
+        {dragOverlayContent}
+        {/* <div style={{ transform: 'translateX(-100%)' }} dangerouslySetInnerHTML={{__html: dragOverlayContent}}></div> */}
+      </DragOverlay>
+      <DisplayedMapProvider>
+        <div className={cls('nb-action-bar', `align-${align}`)}>
+          <div style={{ width: '50%' }}>
+            <Actions align={'left'} />
+          </div>
+          <div style={{ marginLeft: 'auto', width: '50%', textAlign: 'right' }}>
+            <Actions align={'right'} />
+          </div>
+          <AddActionButton />
+        </div>
+      </DisplayedMapProvider>
+    </DndContext>
+  );
+});
