+++ packages/client/src/schemas/table/TableSortHandle.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+import { observer } from '@formily/react';
+import { SortableRowHandle } from './Sortable';
+
+export const TableSortHandle = observer((props: any) => {
+  return <SortableRowHandle {...props} />;
+});
