+++ packages/client/src/schemas/table/hooks/useDefaultRowSelection.ts
@@ -1,0 +1,5 @@
+export const useDefaultRowSelection = () => {
+  return {
+    type: 'checkbox',
+  };
+};
