+++ packages/client/src/schemas/table/hooks/useDefaultSelectedRowKeys.ts
@@ -1,0 +1,6 @@
+import { useState } from "react";
+
+export const useDefaultSelectedRowKeys = () => {
+  const [selectedRowKeys, setSelectedRowKeys] = useState<any>([]);
+  return { selectedRowKeys, setSelectedRowKeys };
+};
