+++ packages/client/src/schemas/table/hooks/useTable.ts
@@ -1,0 +1,6 @@
+import React, { useContext,createContext, useEffect, useState } from 'react';
+import {TableContext} from '../context';
+
+export const useTable = () => {
+  return useContext(TableContext);
+};
