+++ packages/client/src/schemas/table/hooks/useTableActionBars.ts
@@ -1,0 +1,18 @@
+import { useTable } from './useTable';
+
+export const useTableActionBars = () => {
+  const {
+    field,
+    schema,
+    props: { rowKey },
+  } = useTable();
+
+  const bars = schema.reduceProperties((bars, current) => {
+    if (current['x-component'] === 'Table.ActionBar') {
+      return [...bars, current];
+    }
+    return [...bars];
+  }, []);
+
+  return bars;
+};
