+++ packages/client/src/schemas/table/hooks/useTableCreateAction.ts
@@ -1,0 +1,24 @@
+import { useForm } from '@formily/react';
+import { useTable } from './useTable';
+
+export function useTableCreateAction() {
+  const {
+    field,
+    service,
+    resource,
+    refresh,
+    props: { refreshRequestOnChange },
+  } = useTable();
+  const form = useForm();
+  return {
+    async run() {
+      console.log('refreshRequestOnChange', refreshRequestOnChange);
+      if (refreshRequestOnChange) {
+        await resource.create(form.values);
+        await form.reset();
+        return service.refresh();
+      }
+      field.unshift(form.values);
+    },
+  };
+}
