+++ packages/client/src/schemas/table/hooks/useTableRow.ts
@@ -1,0 +1,7 @@
+import React, { useContext } from 'react';
+import {TableRowContext} from '../context';
+
+
+export const useTableRow = () => {
+  return useContext(TableRowContext);
+};
