+++ packages/client/src/schemas/table/hooks/useTableRowRecord.ts
@@ -1,0 +1,7 @@
+import { useContext } from "react";
+import { TableRowContext } from "../context";
+
+export const useTableRowRecord = () => {
+  const ctx = useContext(TableRowContext);
+  return ctx.record;
+};
