+++ packages/client/src/schemas/table/hooks/useTotal.ts
@@ -1,0 +1,10 @@
+import { useTable } from './useTable';
+
+export const useTotal = () => {
+  const {
+    field,
+    service,
+    props: { clientSidePagination },
+  } = useTable();
+  return clientSidePagination ? field?.value?.length : service?.data?.total;
+};
